basis requests completed

+++ server/src/main/java/me/reclaite/bananosbackend/model/report/Report.java
@@ -5,7 +5,6 @@ import jakarta.persistence.Id;
 import lombok.Data;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import me.reclaite.bananosbackend.model.user.User;
 
 @Data
 @Getter
@@ -14,10 +13,10 @@ import me.reclaite.bananosbackend.model.user.User;
 public class Report {
 
     @Id
-    private int id;
+    private Long id;
 
     private ReportType reportType;
     private String description;
-    private int reporterId;
+    private Long reporterId;
 
 }
